requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
torch==2.1.0
torchvision==0.15.2
Pillow==9.5.0
python-dotenv==1.0.0
